Restructure tests to reflect package structure
