+++ Hallo.Test/AspNetCore/Mvc/Supporting/ContactLookup.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Hallo.Test.Integration.TestApi
+namespace Hallo.Test.AspNetCore.Mvc.Supporting
 {
     public class ContactLookup
     {
