+++ Hallo.Test/AspNetCore/Mvc/Supporting/PersonHypermedia.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Hallo.Test.Integration.TestApi
+namespace Hallo.Test.AspNetCore.Mvc.Supporting
 {
     public class PersonRepresentation : Hal<Person>,
                                         IHalLinks<Person>,
